Separate annotations from types. Factilitate Morphs merging, implement default `set_defaults` method
